Extract the custom version of text field from test and use that
