+++ flow-tests/test-frontend/vite-test-assets/pom.xml
@@ -61,6 +61,11 @@
         </dependency>
         <dependency>
             <groupId>com.vaadin</groupId>
+            <artifactId>flow-test-resources</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.vaadin</groupId>
             <artifactId>addon-with-templates</artifactId>
             <version>${project.version}</version>
         </dependency>
