+++ flow-tests/vaadin-spring-tests/pom.xml
@@ -16,6 +16,7 @@
     <properties>
         <maven.deploy.skip>true</maven.deploy.skip>
         <maven.failsafe.plugin.version>3.0.0-M1</maven.failsafe.plugin.version>
+        <vaadin.version>23.0-SNAPSHOT</vaadin.version>
     </properties>
 
     <dependencyManagement>
